+++ src/modules/template/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod check;
 pub mod create;
 pub mod generate;
 pub mod templatespecification;