+++ src/modules/template/src/templatespecification/infrastructure/mod.rs
@@ -1,3 +1,3 @@
 pub mod configuration_loader;
 pub mod folder_loader;
-pub mod liquid_template_renderer;
+pub mod stringreplace_templaterenderer;